[Add] 채팅창 입력
상태관리를 위해서 ChatFooter 컴포넌트를 따로 빼고 상태관리를 함
리랜더링 문제때문
